plugins/ai-mcp-open-responses/skills/agents-sdk-open-responses/scripts/requirements.txt
# OpenAI Agents SDK Dependencies

# Core SDK
openai-agents>=0.1.0
openai>=1.10.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# CLI
rich>=13.7.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
